15052024 aktualizacja szablonów blade,stworzenie nowych szablonów i routingów

+++ Blog/routes/web.php
@@ -25,9 +25,13 @@ Route::middleware('auth')->group(function () {
 Route::post('/posts', [PostController::class, 'store'])->name('posts.store')->middleware('auth');
 Route::get('/posts/create', [PostController::class, 'create'])->name('posts.create')->middleware('auth');
 Route::get('/posts/{post}', [PostController::class, 'show'])->name('posts.show')->middleware('auth');
-Route::delete('/posts/{post}', [PostController::class, 'delete'])->name('posts.delete');
+Route::delete('/posts/{post}', [PostController::class, 'delete'])->name('posts.delete')->middleware('auth');
+Route::get('/posts/{post}/edit', [PostController::class, 'edit'])->name('posts.edit')->middleware('auth');
+Route::put('/posts/{post}', [PostController::class, 'update'])->name('posts.update')->middleware('auth');
 
 //Profile
 Route::get('/profile/{user}', [UserController::class, 'profile_user'])->name('profile');
+Route::get('/manage-photo', [UserController::class, 'manage'])->name('photo')->middleware('auth');
+Route::post('/manage-photo', [UserController::class, 'store_photo'])->name('store.photo')->middleware('auth');
 
 require __DIR__.'/auth.php';
